+++ demisto_sdk/commands/create_id_set/tests/create_id_set_test.py
@@ -1,0 +1,239 @@
+import json
+import os
+import shutil
+
+from collections import OrderedDict
+from tempfile import mkdtemp
+
+from demisto_sdk.commands.common.legacy_git_tools import git_path
+from demisto_sdk.commands.common.update_id_set import ID_SET_ENTITIES
+from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
+from TestSuite.test_tools import ChangeCWD
+from TestSuite.utils import IsEqualFunctions
+
+TESTS_DIR = f'{git_path()}/demisto_sdk/tests'
+
+
+class TestIDSetCreator:
+    def setup(self):
+        self.id_set_full_path = os.path.join(TESTS_DIR, 'test_files', 'content_repo_example', 'id_set.json')
+        self._test_dir = mkdtemp()
+        self.file_path = os.path.join(self._test_dir, 'id_set.json')
+
+    def teardown(self):
+        # delete the id set file
+        try:
+            if os.path.isfile(self.file_path) or os.path.islink(self.file_path):
+                os.unlink(self.file_path)
+            elif os.path.isdir(self.file_path):
+                shutil.rmtree(self.file_path)
+        except Exception as err:
+            print(f'Failed to delete {self.file_path}. Reason: {err}')
+
+    def test_create_id_set_output(self):
+        id_set_creator = IDSetCreator(self.file_path)
+
+        id_set_creator.create_id_set()
+        assert os.path.exists(self.file_path)
+
+    def test_create_id_set_on_specific_pack_output(self):
+        """
+        Given
+        - input - specific pack to create from it ID set
+        - output - path to return the created ID set
+
+        When
+        - create ID set on this pack
+
+        Then
+        - ensure that the created ID set is in the path of the output
+
+        """
+        id_set_creator = IDSetCreator(self.file_path, input='Packs/AMP')
+
+        id_set_creator.create_id_set()
+        assert os.path.exists(self.file_path)
+
+    def test_create_id_set_no_output(self, mocker):
+        import demisto_sdk.commands.common.update_id_set as uis
+        mocker.patch.object(uis, 'cpu_count', return_value=1)
+        id_set_creator = IDSetCreator(output='')
+
+        id_set = id_set_creator.create_id_set()
+        assert not os.path.exists(self.file_path)
+        assert id_set is not None
+        assert 'scripts' in id_set.keys()
+        assert 'integrations' in id_set.keys()
+        assert 'playbooks' in id_set.keys()
+        assert 'TestPlaybooks' in id_set.keys()
+        assert 'Classifiers' in id_set.keys()
+        assert 'Dashboards' in id_set.keys()
+        assert 'IncidentFields' in id_set.keys()
+        assert 'IncidentTypes' in id_set.keys()
+        assert 'IndicatorFields' in id_set.keys()
+        assert 'IndicatorTypes' in id_set.keys()
+        assert 'Layouts' in id_set.keys()
+        assert 'Reports' in id_set.keys()
+        assert 'Widgets' in id_set.keys()
+        assert 'Mappers' in id_set.keys()
+        assert 'Packs' in id_set.keys()
+
+    def test_create_id_set_on_specific_pack(self, repo):
+        """
+        Given
+        - two packs with integrations to create an ID set from
+
+        When
+        - create ID set on one of the packs
+
+        Then
+        - ensure there is only one integration in the ID set integrations list
+        - ensure output id_set contains only the pack on which created the ID set on
+        - ensure output id_set does not contain the second pack
+
+        """
+        packs = repo.packs
+
+        pack_to_create_id_set_on = repo.create_pack('pack_to_create_id_set_on')
+        pack_to_create_id_set_on.create_integration(yml={'commonfields': {'id': 'id1'}, 'category': '', 'name':
+                                                         'integration to create id set', 'script': {'type': 'python'}},
+                                                    name='integration1')
+        packs.append(pack_to_create_id_set_on)
+
+        pack_to_not_create_id_set_on = repo.create_pack('pack_to_not_create_id_set_on')
+        pack_to_not_create_id_set_on.create_integration(yml={'commonfields': {'id2': 'id'}, 'category': '', 'name':
+                                                             'integration to not create id set'}, name='integration2')
+        packs.append(pack_to_not_create_id_set_on)
+
+        id_set_creator = IDSetCreator(self.file_path, pack_to_create_id_set_on.path)
+
+        id_set_creator.create_id_set()
+
+        with open(self.file_path, 'r') as id_set_file:
+            private_id_set = json.load(id_set_file)
+
+        assert len(private_id_set['integrations']) == 1
+        assert private_id_set['integrations'][0].get('id1', {}).get('name', '') == 'integration to create id set'
+        assert private_id_set['integrations'][0].get('id2', {}).get('name', '') == ''
+
+    def test_create_id_set_on_specific_empty_pack(self, repo):
+        """
+        Given
+        - an empty pack to create from it ID set
+
+        When
+        - create ID set on this pack
+
+        Then
+        - ensure that an ID set is created and no error is returned
+        - ensure output id_set is empty
+
+        """
+        pack = repo.create_pack()
+
+        id_set_creator = IDSetCreator(self.file_path, pack.path)
+
+        id_set_creator.create_id_set()
+
+        with open(self.file_path, 'r') as id_set_file:
+            private_id_set = json.load(id_set_file)
+        for content_entity, content_entity_value_list in private_id_set.items():
+            if content_entity != 'Packs':
+                assert len(content_entity_value_list) == 0
+            else:
+                assert len(content_entity_value_list) == 1
+
+
+def test_create_id_set_flow(repo, mocker):
+    # Note: if DEMISTO_SDK_ID_SET_REFRESH_INTERVAL is set it can fail the test
+    mocker.patch.dict(os.environ, {'DEMISTO_SDK_ID_SET_REFRESH_INTERVAL': '-1'})
+    number_of_packs_to_create = 10
+    repo.setup_content_repo(number_of_packs_to_create)
+
+    with ChangeCWD(repo.path):
+        id_set_creator = IDSetCreator(repo.id_set.path, print_logs=False)
+        id_set_creator.create_id_set()
+
+    id_set_content = repo.id_set.read_json_as_dict()
+    assert not IsEqualFunctions.is_dicts_equal(id_set_content, {})
+    assert IsEqualFunctions.is_lists_equal(list(id_set_content.keys()), ID_SET_ENTITIES + ['Packs'])
+    for id_set_entity in ID_SET_ENTITIES:
+        entity_content_in_id_set = id_set_content.get(id_set_entity)
+        assert entity_content_in_id_set
+
+        # Since Layouts folder contains both layouts and layoutcontainers then this folder has 2 * amount objects
+        # And since there is a test playbook for each integration and script.
+        if id_set_entity not in {'Layouts', 'TestPlaybooks'}:
+            assert len(entity_content_in_id_set) == number_of_packs_to_create
+        else:
+            assert len(entity_content_in_id_set) == number_of_packs_to_create * 2
+
+
+integration1 = {
+    'Integration1': OrderedDict([('name', 'Integration1'), ('commands', ['test-command_1', 'test-command'])])}
+integration2 = {
+    'Integration2': OrderedDict([('name', 'Integration2'), ('commands', ['test-command', 'test-command_2'])])}
+expected_output_map = {'test-command': ['Integration1', 'Integration2'],
+                       'test-command_1': ['Integration1'],
+                       'test-command_2': ['Integration2']}
+playbook1 = {
+    'Playbook1': OrderedDict(
+        [('name', 'Playbook1'), ('command_to_integration', {'test-command': "", 'test-command_1': ""})])}
+playbook2 = {
+    'Playbook2': OrderedDict([('name', 'Playbook2'), ('command_to_integration', {'test-command': "",
+                                                                                 'test-command_2': ""})])}
+
+
+def test_create_command_to_implemented_integration_map(repo):
+    """
+
+    Given
+        - a list of integrations
+
+    When
+        - create_command_to_implemented_integration_map is called
+
+    Then
+        - Validates that a dictionary between command name and list of all integration that implement this command
+        was returned.
+
+    """
+    pack = repo.create_pack("Pack1")
+    integration_list = [integration1, integration2]
+
+    id_set_creator = IDSetCreator(pack.path, print_logs=False)
+    id_set_creator.id_set["integrations"] = integration_list
+
+    command_to_implemented_integration_map = id_set_creator.create_command_to_implemented_integration_map()
+    assert command_to_implemented_integration_map == expected_output_map
+
+
+def test_add_command_to_implementing_integrations_mapping(repo):
+    """
+
+    Given
+        - an id_set file includes integrations and playbooks
+
+    When
+        - modify_id_set_command_to_integration_of_playbook is called
+
+    Then
+        - Validates that each command_to_integration in playbook is a dictionary between command name and list of all
+        integration that implement this command.
+
+    """
+    pack = repo.create_pack("Pack1")
+    integration_list = [integration1, integration2]
+    playbook_list = [playbook1, playbook2]
+
+    id_set_creator = IDSetCreator(pack.path, print_logs=False)
+    id_set_creator.id_set["integrations"] = integration_list
+    id_set_creator.id_set["playbooks"] = playbook_list
+
+    id_set_creator.add_command_to_implementing_integrations_mapping()
+
+    playbook_set = id_set_creator.id_set["playbooks"]
+    assert playbook_set[0]["Playbook1"]['command_to_integration']['test-command'] == ['Integration1', 'Integration2']
+    assert playbook_set[0]["Playbook1"]['command_to_integration']['test-command_1'] == ['Integration1']
+    assert playbook_set[1]["Playbook2"]['command_to_integration']['test-command'] == ['Integration1', 'Integration2']
+    assert playbook_set[1]["Playbook2"]['command_to_integration']['test-command_2'] == ['Integration2']
